requirements.txt
android_backup==0.2.0
appdirs==1.4.4
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
construct==2.10.70
croniter==2.0.5
cryptography==42.0.7
defusedxml==0.7.1
idna==3.7
ifaddr==0.2.0
micloud==0.6
netifaces==0.11.0
pycparser==2.22
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
python-miio==0.5.12
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.5.1
six==1.16.0
tqdm==4.66.4
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wheel==0.43.0
zeroconf==0.132.2
